refactor(models): deduplicate MoLeRGenerator and MoLeRVae by adding common MoLeRBaseModel
