+++ molecule_generation/models/moler_vae.py
@@ -1,6 +1,7 @@
 """MoLeR Variational Autoencoder model."""
 from dataclasses import dataclass
-from typing import Any, Dict, Optional, Tuple, List, NamedTuple, Union, Iterable
+from typing import Any, Dict, Optional, Tuple, List, NamedTuple, Union
+from molecule_generation.models.moler_base_model import MoLeRBaseModel
 
 import numpy as np
 import tensorflow as tf
@@ -10,18 +11,16 @@ from tf2_gnn.layers import NodesToGraphRepresentationInput, WeightedSumGraphRepr
 
 from molecule_generation.dataset.trace_dataset import TraceDataset
 from molecule_generation.layers.moler_decoder import (
-    MoLeRDecoder,
     MoLeRDecoderInput,
     MoLeRDecoderMetrics,
+    MoLeRDecoderOutput,
 )
-from molecule_generation.utils.training_utils import get_class_balancing_weights
 from molecule_generation.utils.property_models import (
     PropertyTaskType,
     PropertyPredictionLayer,
     MLPBinaryClassifierLayer,
     MLPRegressionLayer,
 )
-from molecule_generation.utils.epoch_metrics_logger import EpochMetricsLogger
 
 
 @dataclass
@@ -47,7 +46,7 @@ class PropertyPredictionMetrics(NamedTuple):
     property_to_metrics: Dict[str, Any]
 
 
-class MoLeRVae(GraphTaskModel):
+class MoLeRVae(MoLeRBaseModel):
     """[Mo]lecule [Le]vel [R]representation-based VAE model for handling molecular graphs.
 
     The components of our model work together as follows:
@@ -57,17 +56,10 @@ class MoLeRVae(GraphTaskModel):
     The vocabulary of motifs has to be precomputed during preprocessing.
     """
 
-    __decoder_prefix = "decoder_"
-
-    @classmethod
-    def decoder_prefix(cls) -> str:
-        return cls.__decoder_prefix
-
     @classmethod
     def get_default_hyperparameters(cls, mp_style: Optional[str] = None) -> Dict[str, Any]:
         base_hypers = super().get_default_hyperparameters(mp_style)
 
-        # Add our own hyperparameters:
         base_hypers.update(
             {
                 # Encoder GNN hyperparameters:
@@ -80,7 +72,6 @@ class MoLeRVae(GraphTaskModel):
                 "use_intermediate_gnn_results": True,  # Use all intermediate results of the encoder GNN
                 "categorical_features_embedding_dim": 64,
                 # Encoder -> Graph-level representation hyperparameters:
-                "latent_repr_size": 512,
                 "latent_repr_mlp_layers": [512, 512],
                 "latent_repr_num_heads": 32,
                 "latent_repr_dropout_rate": 0.0,
@@ -88,49 +79,20 @@ class MoLeRVae(GraphTaskModel):
                 # Property prediction hyperparameters:
                 "property_predictor_mlp_layers": [64, 32],
                 "property_predictor_mlp_dropout_rate": 0.0,
-                # Relative contributions of the different next step prediction losses:
-                "node_classification_loss_weight": 1.0,
-                "first_node_classification_loss_weight": 0.07,
-                "edge_selection_loss_weight": 1.0,
-                "edge_type_loss_weight": 1.0,
-                "attachment_point_selection_weight": 1.0,
                 # Relative contribution of the KL loss term. This was tuned for a motif
                 # vocabulary of 128, may need to be adjusted for larger vocabularies
                 # (empirically increments of 0.005 every time vocabulary size doubles):
                 "kl_divergence_weight": 0.02,
                 "kl_divergence_annealing_beta": 0.999,  # 10% weight after 100 steps, 63% after 1000, 99% after 5000
-                # Training hyperparameters:
-                "learning_rate": 0.001,
-                "gradient_clip_value": 0.5,
-                "num_train_steps_between_valid": 0,  # Default value of 0 means that whole dataset will be run through.
-                # Not really hyperparameters, but logging parameters:
-                "logged_loss_smoothing_window_size": 100,  # Number of training batches to include in the reported moving average training loss.
             }
         )
-
-        # Add hyperparameters for the GNN decoder:
-        decoder_hypers = {
-            cls.decoder_prefix() + k: v
-            for k, v in MoLeRDecoder.get_default_params(mp_style).items()
-        }
-        base_hypers.update(decoder_hypers)
 
         return base_hypers
 
     def __init__(self, params: Dict[str, Any], dataset: TraceDataset, **kwargs):
         super().__init__(params, dataset, **kwargs)
 
-        # Shortcuts to commonly used hyperparameters:
-        self._latent_repr_dim = params["latent_repr_size"]
         self._latent_sample_strategy = params["latent_sample_strategy"]
-
-        # Keep track of the training step as a TF variable
-        self._train_step_counter = tf.Variable(
-            initial_value=0, trainable=False, dtype=tf.int32, name="training_step"
-        )
-
-        # Get some information out from the dataset:
-        next_node_type_distribution = dataset.metadata.get("train_next_node_type_distribution")
 
         # ===== Prepare sub-layers, which will be actually created in .build().
         # Layer from per-node encoder GNN results to per-graph representation:
@@ -195,79 +157,6 @@ class MoLeRVae(GraphTaskModel):
             else:
                 raise ValueError(f"Unknown property type {prop_type}")
 
-        class_weight_factor = self._params.get("node_type_predictor_class_loss_weight_factor", 0.0)
-
-        if not (0.0 <= class_weight_factor <= 1.0):
-            raise ValueError(
-                f"Node class loss weight node_classifier_class_loss_weight_factor must be in [0,1], but is {class_weight_factor}!"
-            )
-
-        if class_weight_factor > 0:
-            atom_type_nums = [
-                next_node_type_distribution[dataset.node_type_index_to_string[type_idx]]
-                for type_idx in range(dataset.num_node_types)
-            ]
-            atom_type_nums.append(next_node_type_distribution["None"])
-
-            class_weights = get_class_balancing_weights(
-                class_counts=atom_type_nums, class_weight_factor=class_weight_factor
-            )
-        else:
-            class_weights = None
-
-        motif_vocabulary = dataset.metadata.get("motif_vocabulary")
-        self._uses_motifs = motif_vocabulary is not None
-
-        self._node_categorical_num_classes = dataset.node_categorical_num_classes
-
-        if self.uses_categorical_features:
-            if "categorical_features_embedding_dim" in self._params:
-                self._node_categorical_features_embedding = None
-            else:
-                # Older models use one hot vectors instead of dense embeddings, simulate that here.
-                self._params[
-                    "categorical_features_embedding_dim"
-                ] = self._node_categorical_num_classes
-                self._node_categorical_features_embedding = np.eye(
-                    self._node_categorical_num_classes, dtype=np.float32
-                )
-
-        # Finally, the decoder layer, which does all kinds of important things:
-        decoder_prefix = self.decoder_prefix()
-        n = len(decoder_prefix)
-        decoder_hypers = {k[n:]: v for k, v in self._params.items() if k.startswith(decoder_prefix)}
-        self._decoder_layer = MoLeRDecoder(
-            decoder_hypers,
-            name="MoLeRDecoder",
-            atom_featurisers=dataset.metadata.get("feature_extractors", []),
-            index_to_node_type_map=dataset.node_type_index_to_string,
-            node_type_loss_weights=class_weights,
-            motif_vocabulary=motif_vocabulary,
-            node_categorical_num_classes=self._node_categorical_num_classes,
-        )
-
-        # Moving average variables, will be filled later:
-        self._logged_loss_smoothing_window_size = self._params["logged_loss_smoothing_window_size"]
-
-        # Deal with Tensorboard's global state:
-        tf.summary.experimental.set_step(0)
-
-    @property
-    def latent_dim(self):
-        return self._latent_repr_dim
-
-    @property
-    def decoder(self):
-        return self._decoder_layer
-
-    @property
-    def uses_motifs(self) -> bool:
-        return self._uses_motifs
-
-    @property
-    def uses_categorical_features(self) -> bool:
-        return self._node_categorical_num_classes is not None
-
     def build(self, input_shapes: Dict[str, Any]):
         # Compute some sizes and shapes we'll re-use a few times:
         final_node_representation_dim = self._params["gnn_hidden_dim"]
@@ -295,28 +184,7 @@ class MoLeRVae(GraphTaskModel):
             with tf.name_scope(f"property_{prop_name}"):
                 prop_predictor.build(latent_graph_representation_shape)
 
-        with tf.name_scope("decoder"):
-            partial_adjacency_lists: Tuple[tf.TensorShape, ...] = tuple(
-                input_shapes[f"partial_adjacency_list_{edge_type_idx}"]
-                for edge_type_idx in range(self._num_edge_types)
-            )
-            self._decoder_layer.build(
-                MoLeRDecoderInput(
-                    node_features=tf.TensorShape((None, input_shapes["partial_node_features"][-1])),
-                    node_categorical_features=tf.TensorShape((None,)),
-                    adjacency_lists=partial_adjacency_lists,
-                    num_graphs_in_batch=input_shapes["num_partial_graphs_in_batch"],
-                    graph_to_focus_node_map=input_shapes["focus_nodes"],
-                    node_to_graph_map=input_shapes["node_to_partial_graph_map"],
-                    input_molecule_representations=latent_graph_representation_shape,
-                    graphs_requiring_node_choices=input_shapes[
-                        "partial_graphs_requiring_node_choices"
-                    ],
-                    candidate_edges=input_shapes["valid_edge_choices"],
-                    candidate_edge_features=input_shapes["edge_features"],
-                    candidate_attachment_points=input_shapes["valid_attachment_point_choices"],
-                )
-            )
+        self._build_decoder(input_shapes=input_shapes)
 
         if self.uses_categorical_features and self._node_categorical_features_embedding is None:
             with tf.name_scope("node_categorical_features_embedding"):
@@ -329,7 +197,7 @@ class MoLeRVae(GraphTaskModel):
                     trainable=True,
                 )
 
-        super().build(input_shapes)
+        GraphTaskModel.build(self, input_shapes)
 
     def get_initial_node_feature_shape(self, input_shapes) -> tf.TensorShape:
         node_features_shape = super().get_initial_node_feature_shape(input_shapes)
@@ -430,6 +298,7 @@ class MoLeRVae(GraphTaskModel):
         graph_mean = graph_mean_and_log_variance[:, : self.latent_dim]  # Shape: [V, MD]
         graph_log_variance = graph_mean_and_log_variance[:, self.latent_dim :]  # Shape: [V, MD]
 
+        # result_representations: shape [PG, MD]
         if self._latent_sample_strategy == "passthrough":
             result_representations = tf.gather(graph_mean, partial_graph_to_original_graph_map)
         elif self._latent_sample_strategy == "per_graph":
@@ -477,12 +346,7 @@ class MoLeRVae(GraphTaskModel):
             batch_features[f"partial_adjacency_list_{edge_type_idx}"]
             for edge_type_idx in range(self._num_edge_types)
         )  # Each element has shape (E, 2)
-        (
-            node_type_logits,
-            edge_candidate_logits,
-            edge_type_logits,
-            attachment_point_selection_logits,
-        ) = self._decoder_layer(
+        decoder_output: MoLeRDecoderOutput = self._decoder_layer(
             MoLeRDecoderInput(
                 node_features=batch_features["partial_node_features"],
                 node_categorical_features=batch_features["partial_node_categorical_features"],
@@ -529,11 +393,11 @@ class MoLeRVae(GraphTaskModel):
         return MoLeRVaeOutput(
             graph_representation_mean=graph_representation_mean,
             graph_representation_log_variance=graph_representation_log_variance,
-            node_type_logits=node_type_logits,
+            node_type_logits=decoder_output.node_type_logits,
             first_node_type_logits=first_node_type_logits,
-            edge_candidate_logits=edge_candidate_logits,
-            edge_type_logits=edge_type_logits,
-            attachment_point_selection_logits=attachment_point_selection_logits,
+            edge_candidate_logits=decoder_output.edge_candidate_logits,
+            edge_type_logits=decoder_output.edge_type_logits,
+            attachment_point_selection_logits=decoder_output.attachment_point_selection_logits,
             predicted_properties=property_prediction_results,
         )
 
@@ -568,8 +432,8 @@ class MoLeRVae(GraphTaskModel):
         batch_labels: Dict[str, tf.Tensor],
     ) -> MoLeRMetrics:
 
-        decoder_metrics = self.decoder.compute_metrics(
-            batch_features=batch_features, batch_labels=batch_labels, task_output=task_output
+        total_loss, decoder_metrics = self._compute_decoder_loss_and_metrics(
+            batch_features=batch_features, task_output=task_output, batch_labels=batch_labels
         )
 
         kl_divergence_summand = (
@@ -592,21 +456,7 @@ class MoLeRVae(GraphTaskModel):
             batch_features["num_graphs_in_batch"], tf.float32
         )
 
-        total_loss = (
-            kl_div_weight * normalised_kl_loss
-            + self._params["node_classification_loss_weight"]
-            * decoder_metrics.node_classification_loss
-            + self._params["first_node_classification_loss_weight"]
-            * decoder_metrics.first_node_classification_loss
-            + self._params["edge_selection_loss_weight"] * decoder_metrics.edge_loss
-            + self._params["edge_type_loss_weight"] * decoder_metrics.edge_type_loss
-        )
-
-        if self.uses_motifs:
-            total_loss += (
-                self._params["attachment_point_selection_weight"]
-                * decoder_metrics.attachment_point_selection_loss
-            )
+        total_loss += kl_div_weight * normalised_kl_loss
 
         # Don't use 'asdict': that tries to deepcopy decoder_metrics and fails with TypeError
         return MoLeRMetrics(
@@ -636,24 +486,13 @@ class MoLeRVae(GraphTaskModel):
 
         return PropertyPredictionMetrics(total_loss, property_to_metrics)
 
+    def _get_graph_generation_losses(self, task_results: List[Any]) -> List[Tuple[str, str]]:
+        graph_generation_losses = super()._get_graph_generation_losses(task_results)
+        average_kl_divergence = self._dict_average(task_results, "kl_divergence")
+        graph_generation_losses.append(("Avg KL divergence:", f"{average_kl_divergence: 7.4f}\n"))
+        return graph_generation_losses
+
     def compute_epoch_metrics(self, task_results: List[Any]) -> Tuple[float, str]:
-        def dict_average(key):
-            return np.average([r[key] for r in task_results])
-
-        average_loss = dict_average("loss")
-        average_node_classification_loss = dict_average("node_classification_loss")
-        average_first_node_classification_loss = dict_average("first_node_classification_loss")
-        average_edge_loss = dict_average("edge_loss")
-        average_edge_type_loss = dict_average("edge_type_loss")
-
-        if self.uses_motifs:
-            average_attachment_point_selection_loss = dict_average(
-                "attachment_point_selection_loss"
-            )
-        else:
-            average_attachment_point_selection_loss = None
-
-        average_kl_divergence = dict_average("kl_divergence")
 
         # Compute results for the individual property predictors.
         # Weigh their respective contributions using the loss weight.
@@ -677,58 +516,16 @@ class MoLeRVae(GraphTaskModel):
             )
             property_epoch_descriptions.append(f"{prop_name}: {prop_metric_description}")
 
-        graph_generation_losses = [
-            ("Avg node class. loss:", f"{average_node_classification_loss: 7.4f}\n"),
-            ("Avg first node class. loss:", f"{average_first_node_classification_loss: 7.4f}\n"),
-            ("Avg edge selection loss:", f"{average_edge_loss: 7.4f}\n"),
-            ("Avg edge type loss:", f"{average_edge_type_loss: 7.4f}\n"),
-        ]
-
-        if self.uses_motifs:
-            graph_generation_losses.append(
-                (
-                    "Avg attachment point selection loss:",
-                    f"{average_attachment_point_selection_loss: 7.4f}\n",
-                )
-            )
-
-        graph_generation_losses.append(("Avg KL divergence:", f"{average_kl_divergence: 7.4f}\n"))
-
+        average_loss = self._dict_average(task_results, "loss")
         result_string = (
             f"\n"
             f"Avg weighted sum. of graph losses: {average_loss: 7.4f}\n"
             f"Avg weighted sum. of prop losses:  {property_epoch_metric: 7.4f}\n"
         )
 
-        name_column_width = max(len(prefix) for (prefix, _) in graph_generation_losses) + 1
-
-        for prefix, loss in graph_generation_losses:
-            # Use extra spaces to allign all graph generation losses.
-            result_string += prefix.ljust(name_column_width) + loss
+        graph_generation_losses = self._get_graph_generation_losses(task_results)
+        result_string += self._format_graph_generation_losses(graph_generation_losses)
 
         result_string += f"Property results: {' | '.join(property_epoch_descriptions)}"
 
         return average_loss + property_epoch_metric, result_string
-
-    def run_on_data_iterator(
-        self,
-        data_iterator: Iterable[Tuple[Dict[str, tf.Tensor], Dict[str, tf.Tensor]]],
-        quiet: bool = False,
-        training: bool = True,
-        max_num_steps: Optional[int] = None,  # Run until dataset ends if None
-        aml_run: Optional = None,
-    ) -> Tuple[float, float, List[Any]]:
-        with EpochMetricsLogger(
-            window_size=self._logged_loss_smoothing_window_size,
-            quiet=quiet,
-            aml_run=aml_run,
-            training=training,
-        ) as metrics_logger:
-            for step, (batch_features, batch_labels) in enumerate(data_iterator):
-                if max_num_steps and step >= max_num_steps:
-                    break
-
-                task_metrics = self._run_step(batch_features, batch_labels, training)
-                metrics_logger.log_step_metrics(task_metrics, batch_features)
-
-        return metrics_logger.get_epoch_summary()
